+++ src/Models/Subscription.php
@@ -23,6 +23,7 @@ class Subscription extends Model
     protected $casts = [
         'next_payment_due_date' => 'datetime',
         'metadata' => AsArrayObject::class,
+        'quantity' => 'integer',
     ];
 
     public function getTable(): string